Add clarification that functions are not supported
I did run into this issue as I was trying to use the length of an array is the `inputPath`, resulting the in the following error.

```
An error occurred while executing the state 'MyState' (entered at the event id #4). Invalid path '$.list.length()' : Property ['length()'] not found in path $['list']
```

Even though the current documentation technically covers it by stating:

>  …that it can identify only a single node in…

I thought it may still be helpful to make it explicit that functions as defined in the JsonPath spec are not supported at all.