Remove overly-broad permission example
We shouldn't be encouraging the use of `*` when it's not absolutely required. In this case, executions have defined ARNs that can be specified more narrowly than `*`, so the documentation should follow best-practices and show how to make the permission more narrow.

In an attempt to mirror the style of the rest of this documentation, this ARN format hardcodes the partition to `aws`, though that partition is also variable, like the region. A working snippet of CFN code that would follow this pattern while keeping the partition dynamic would be:

```
"Fn::Sub": "arn:${AWS::Partition}:states:${AWS::Region}:${AWS::AccountId}:execution:${StateMachine.Name}:*"
```